Remove procholders + some extra (#2344)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Includes:
[Replacing procholders with
actions](https://github.com/tgstation/tgstation/pull/67083) (The main
thing)
[Using lazyaccess for params
modifiers](https://github.com/tgstation/tgstation/pull/56846) Because
its good
[Datumized view
range](https://github.com/tgstation/tgstation/pull/51208) Because its
also good
[Can pass refactor](https://github.com/tgstation/tgstation/pull/48659)
for projectiles and its good
[Projectile Piercing
refactor](https://github.com/tgstation/tgstation/pull/54924) For
antimagic
[Area Flags](https://github.com/tgstation/tgstation/pull/52751) for
jaunt and teleporting
[Attack secondary chain added properly for
items](https://github.com/tgstation/tgstation/pull/56663) (and friends)
for touch spells

closes #2322

Hello ladies and gentlemen, to the proc_holder removal pr (and *some*
other stuff of course). This pr replaces all proc_holders with real
actions.

Casting should also be MUCH more consistent (unlike bows which fail even
on a local server with minimum ping), due to being taken off the client
processing system for item charge.

Spell costs have been consolidated so miracle costs aren't seperate and
bugs involving casting like spells ignoring cast range (potentially?)
have been fixed.

TODO
- [x] Change action screenlocs
- [x] Fix hands being inverted (even tho it makes more sense wtf)
- [x] Fix stupid mouse issues
- [ ] Add more fitting antimagic effects

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
It absolutely is

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] Hell yes
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

+++ code/game/objects/items/natural.dm
@@ -28,22 +28,24 @@
 	else
 		return ..()
 
-/obj/item/natural/pre_attack_right(atom/A, mob/living/user, params)
+/obj/item/natural/pre_attack_secondary(atom/A, mob/living/user, params)
+	. = ..()
+	if(. == SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN)
+		return
 	if(istype(A, /obj/item/natural))
 		if(item_flags & IN_STORAGE)
 			to_chat(user, span_warning("It's hard to find [src] in my bag."))
-			return
+			return SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 		var/obj/item/natural/B = A
 		if(bundletype && bundletype == B.bundletype)
 			if(!user.temporarilyRemoveItemFromInventory(src))
-				return TRUE
+				return SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 			var/obj/item/natural/bundle/N = new bundletype(loc)
 			to_chat(user, span_notice("You collect the [N.stackname] into a bundle."))
 			qdel(B)
 			qdel(src)
 			user.put_in_active_hand(N)
-			return TRUE
-	return ..()
+			return SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 
 /obj/item/natural/bundle
 	name = "bundle"
@@ -104,50 +106,56 @@
 	else
 		return ..()
 
-/obj/item/natural/bundle/attack_right(mob/user)
+/obj/item/natural/bundle/attack_hand_secondary(mob/user, params)
+	. = ..()
+	if(. == SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN)
+		return
+	. = SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 	if(item_flags & IN_STORAGE)
+		to_chat(user, span_warning("I can't reach [src]!"))
 		return
 	if(amount <= 0) //how did you manage to do this
 		qdel(src)
 		return
-	var/atom/item = user.get_active_held_item()
-	if(item && item.type != stacktype)
-		return ..()
 	var/mob/living/carbon/human/H = user
 	switch(amount)
 		if(2)
 			if(!user.temporarilyRemoveItemFromInventory(src))
 				return
-			var/obj/F = new stacktype(src.loc)
-			var/obj/I = new stacktype(src.loc)
+			var/obj/F = new stacktype(get_turf(src))
+			new stacktype(get_turf(src))
 			H.put_in_hands(F)
-			H.put_in_hands(I)
 			qdel(src)
 			return
 		if(1)
 			if(!user.temporarilyRemoveItemFromInventory(src))
 				return
-			var/obj/F = new stacktype(src.loc)
+			var/obj/F = new stacktype(get_turf(src))
 			H.put_in_hands(F)
 			qdel(src)
 			return
 		else
 			amount -= 1
-			var/obj/F = new stacktype(src.loc)
+			var/obj/F = new stacktype(get_turf(src))
 			H.put_in_hands(F)
 			to_chat(user, span_notice("You remove \a [F] from [src]."))
+
 	update_bundle()
 
 /obj/item/natural/bundle/examine(mob/user)
 	. = ..()
 	. += span_notice("There are [amount] [stackname] in this bundle.")
 
-/obj/item/natural/bundle/pre_attack_right(atom/A, mob/living/user, params)
+/obj/item/natural/bundle/pre_attack_secondary(atom/A, mob/living/user, params)
+	. = ..()
+	if(. == SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN)
+		return
+	. = SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 	if(amount <= 0) //how did you manage to do this
 		qdel(src)
 		return
 	if(ismob(A))
-		return ..()
+		return SECONDARY_ATTACK_CALL_NORMAL
 	user.changeNext_move(CLICK_CD_MELEE)
 	if(amount >= maxamount)
 		to_chat(user, span_warning("There's not enough space in [src]."))
@@ -181,7 +189,6 @@
 					amount += B.amount
 					qdel(B)
 		update_bundle()
-	return TRUE
 
 /obj/item/natural/bundle/proc/update_bundle()
 	if(firefuel != 0)
