+++ code/game/objects/items/natural/wood.dm
@@ -56,10 +56,9 @@
 	. = ..()
 
 
-/obj/item/grown/log/tree/attack_right(mob/living/user)
+/obj/item/grown/log/tree/attackby_secondary(obj/item/I, mob/living/user, params)
 	. = ..()
-	var/obj/item/I = user.get_active_held_item()
-	if(!I)
+	if(. == SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN)
 		return
 
 	if(user.used_intent.blade_class == BCLASS_CHOP && lumber_amount && lumber_alt)
@@ -68,7 +67,7 @@
 		var/minimum = 1
 		playsound(src, 'sound/misc/woodhit.ogg', 100, TRUE)
 		if(!do_after(user, lumber_time, user))
-			return
+			return SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 		if(skill_level > 0) // If skill level is 1 or higher, we get more minimum wood!
 			minimum = 2
 		lumber_amount = rand(minimum, max(round(skill_level), minimum))
@@ -87,8 +86,7 @@
 		user.mind.add_sleep_experience(/datum/skill/labor/lumberjacking, (user.STAINT*0.5))
 		playsound(src, destroy_sound, 100, TRUE)
 		qdel(src)
-		return TRUE
-	..()
+		return SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 
 /*
 * Okay so the root of this proc defines dissasemble
@@ -175,7 +173,7 @@
 				L.update_sneak_invis(TRUE)
 			L.consider_ambush()
 
-/obj/item/grown/log/tree/stick/attack_self(mob/living/user)
+/obj/item/grown/log/tree/stick/attack_self(mob/living/user, params)
 	user.visible_message("<span class='warning'>[user] snaps [src].</span>")
 	playsound(user,'sound/items/seedextract.ogg', 100, FALSE)
 	qdel(src)
@@ -192,15 +190,18 @@
 		return
 	return ..()
 
-/obj/item/grown/log/tree/stick/attack_right(mob/living/user)
+/obj/item/grown/log/tree/stick/attackby_secondary(obj/item/I, mob/user, params)
 	. = ..()
-	var/obj/item/I = user.get_active_held_item()
+	if(. == SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN)
+		return
+
 	if(istype(I, /obj/item/grown/log/tree/stick))
 		var/obj/item/natural/bundle/stick/F = new(get_turf(user))
 		qdel(I)
 		qdel(src)
 		user.put_in_hands(F)
 		to_chat(user, "You collect the [F.stackname] into a bundle.")
+		return SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 
 /obj/item/grown/log/tree/stake
 	name = "stake"
